+++ node/src/main/kotlin/net/corda/node/services/transactions/NonValidatingNotaryFlow.kt
@@ -4,6 +4,7 @@ import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.flows.NotaryFlow
 import net.corda.core.flows.TransactionParts
 import net.corda.core.identity.Party
+import net.corda.core.internal.datavending.FetchDataFlow
 import net.corda.core.node.services.TrustedAuthorityNotaryService
 import net.corda.core.transactions.FilteredTransaction
 import net.corda.core.utilities.unwrap
@@ -23,6 +24,7 @@ class NonValidatingNotaryFlow(otherSide: Party, service: TrustedAuthorityNotaryS
             it.verify()
             it
         }
+        send(otherSide, FetchDataFlow.EndRequest)
         return TransactionParts(ftx.rootHash, ftx.filteredLeaves.inputs, ftx.filteredLeaves.timeWindow)
     }
 }